+++ cope/src/cope/servlet/board/BoardDeleteServlet.java
@@ -1,4 +1,4 @@
-package cope.board.servlet;
+package cope.servlet.board;
 
 import java.io.IOException;
