+++ cope/src/cope/servlet/board/BoardInsertServlet.java
@@ -1,4 +1,4 @@
-package cope.board.servlet;
+package cope.servlet.board;
 
 import java.io.IOException;
 
@@ -11,24 +11,25 @@ import javax.servlet.http.HttpServletResponse;
 import cope.beans.BoardDao;
 import cope.beans.BoardDto;
 
-@WebServlet (urlPatterns = "/manage/editBoard.kh")
-public class BoardEditServlet extends HttpServlet{
+@WebServlet (urlPatterns = "/manage/createBoard.kh")
+public class BoardInsertServlet extends HttpServlet{
 	@Override
 	protected void doGet(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
 		try {
-			//준비 : 2개 게시판 번호, 바꿀게시판이름
+			
+			//준비 2개(게시판 이름, 상위게시판을 의미 하는 0)
 			req.setCharacterEncoding("UTF-8");
 			BoardDto boardDto = new BoardDto();
-			boardDto.setBoardNo(Integer.parseInt(req.getParameter("boardNo")));
 			boardDto.setBoardName(req.getParameter("boardName"));
-			
+			boardDto.setBoardSuperNo(Integer.parseInt(req.getParameter("boardSuperNo")));
 			
 			//처리
 			BoardDao boardDao = new BoardDao();
-			boardDao.editBoard(boardDto);
+			boardDao.insertBoardSuper(boardDto);
 			
 			//출력
 			resp.sendRedirect("manageBoard.jsp");
+
 			
 		} catch (Exception e) {
 			e.printStackTrace();
